Adding JUnit4 dependency
