+++ pom.xml
@@ -22,7 +22,6 @@
     <hadoop.version>2.7.2</hadoop.version>
     <hive.version>2.3.4</hive.version>
     <jdk.version>1.7</jdk.version>
-    <junit.version>5.4.0</junit.version>
   </properties>
 
   <dependencies>
@@ -98,11 +97,11 @@
       <version>4.12</version>
       <!--scope>test</scope-->
     </dependency>
-    <dependency>
+    <!--dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-engine</artifactId>
       <version>${junit.version}</version>
-    </dependency>
+    </dependency-->
     <!-- below removes "Failed to set setXIncludeAware(true)" warnings in log when running cobertura -->
     <dependency>
       <groupId>xerces</groupId>
